feat: Move devtools into it's own crate
