+++ devtools/Cargo.toml
@@ -1,6 +1,5 @@
 [package]
-name = "freya"
-description = "Cross-platform GUI Library."
+name = "freya-devtools"
 version = "0.1.0"
 edition = "2021"
 license = "MIT"
@@ -11,34 +10,18 @@ repository = "https://github.com/marc2332/freya"
 keywords = ["gui", "ui", "cross-platform", "dioxus", "skia", "graphics"]
 categories = ["GUI"]
 
-[features]
-devtools = ["dep:freya-devtools"]
-
-[target."cfg(target_os = \"linux\")".dependencies.skia-safe]
-version = "0.58.0"
-features = ["gl", "textlayout", "svg", "x11", "wayland"]
-
 [dependencies]
-freya-devtools = { path = "../devtools", version = "0.1.0", optional = true }
 freya-node-state = { path = "../state", version = "0.1.0" }
 freya-layout = { path = "../layout", version = "0.1.0" }
 freya-renderer = { path = "../renderer", version = "0.1.0" }
 freya-elements = { path = "../elements", version = "0.1.0" }
-freya-common = { path = "../common", version = "0.1.0" }
 freya-hooks = { path = "../hooks"}
 freya-core = { path = "../core"}
 freya-components ={ path = "../components"}
 dioxus = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519", features = ["macro", "hooks", "hot-reload"]}
-dioxus-rsx = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
 dioxus-native-core = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
 dioxus-core-macro = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
-dioxus-hooks = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
 dioxus-core = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
 dioxus-router = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
-dioxus-hot-reload = { git = "https://github.com/DioxusLabs/dioxus", rev="49c5a5043a16fc82210af146c345793dd448e519"}
 tokio = { version = "1.23.0", features = ["sync", "rt-multi-thread", "time", "macros"] }
-anymap = "0.12.1"
-fxhash = "0.2.1"
 skia-safe = { version = "0.58.0", features = ["gl", "textlayout", "svg"] }
-tracing = "0.1"
-futures = "0.3.25"
